requirements.txt
aiogram
pydantic-settings
aiosqlite
APScheduler
rapidfuzz
boto3
aiohttp
python-dateutil
pyyaml
psutil